requirements.txt
plotly==5.24.1
plotly-express==0.4.1
streamlit-option-menu==0.4.0
pandas==2.2.3
openpyxl==3.1.5
yfinance==0.2.44